tradefinancecc/go/tradefinancecc_test.go
package main

import (
	"fmt"
	"testing"

	"github.com/hyperledger/fabric/core/chaincode/shim"
)

func checkInit(t *testing.T, stub *shim.MockStub, args [][]byte) {
	res := stub.MockInit("1", args)
	if res.Status != shim.OK {
		fmt.Println("Init failed", string(res.Message))
		t.FailNow()
	}
	fmt.Println("success init")

}

func checkInvoke(t *testing.T, stub *shim.MockStub, args [][]byte) {

	res := stub.MockInvoke("1", args)
	if res.Status != shim.OK {
		fmt.Println("Invoke", args, "failed", string(res.Message))
		t.FailNow()
	}
}
func checkState(t *testing.T, stub *shim.MockStub, name string, value string) {
	bytes := stub.State[name]
	if bytes == nil {
		fmt.Println("State", name, "failed to get value")
		t.FailNow()
	}
	if string(bytes) != value {
		fmt.Println("State value", name, "was not", value, "as expected")
		t.FailNow()
	}
}

func TestTradefinancecc_Init(t *testing.T) {
	scc := new(SimpleAsset)
	stub := shim.NewMockStub("tradefinancecc", scc)

	// Init A=123 B=234
	checkInit(t, stub, [][]byte{[]byte("init"), []byte("a"), []byte("100"), []byte("b"), []byte("200")})

	checkState(t, stub, "a", "100")
	checkState(t, stub, "b", "200")
}
